refactor(T8a): impl-split SFConveyorBeltHologram (spline/mesh -> _Spline.cpp) [BUILD-PASSED]
Slice T8a. Splits the belt hologram across 2 .cpp (one class): spline setup / auto-route /
build-mode routing / spline-data update / continuous-position-correction / mesh-apply move to
SFConveyorBeltHologram_Spline.cpp (552 lines). SFConveyorBeltHologram.cpp 2,220 -> 1,700 (<2k).

Pure impl-split, ZERO behavior change (methods stay ASFConveyorBeltHologram members, just in a
second TU; the class-scoped Friend grant + accessors still apply, no AccessTransformer change).
Deviates from the card's cross-hologram dedup: the belt vs pipe spline code is parallel-but-
DIFFERENT (factory vs pipe connectors, different mesh-gen), so a shared FSFHologramSplineRouter
would be a behavior-sensitive refactor, not a clean move; deferred to the optional T8-MVP. This
realization meets criterion #5 for the belt hologram safely. Shipping PackagePlugin clean
(ExitCode=0). Solo lane (impl-split) -> belt placement light-check folds into the next smoke.
